# docs/v1.openapi.yaml
openapi: 3.1.0
info:
  title: Claims service API
  version: "1.0"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
  responses:
...
    badRequest:
      description: Допущена синтаксическая ошибка в запросе
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
...
paths:
...
  /v1/districts/:
    get:
      description: Получение доступных областей для создания заявок
      tags: 
        - map
      security: 
        - bearerAuth: []
      responses:
        "200":
          description: Области успешно получены
          content:
            application/json:
              schema:
                type: object
                properties:
                  districts:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        polygon:
                          type: array
                          items:
                            type: object
                            properties:
                              longitude:
                                type: number
                                format: float64
                              latitude:
                                type: number
                                format: float64
  /v1/districts/check-point:
    get:
      description: Проверка нахождения точки в одной из доступных областей
      tags: 
        - map
      security: 
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: latitude
        - name: long
          in: query
          required: true
          schema:
            type: number
            format: longitude
      responses:
        "400":
          description: Один из параметров имеет неверный формат, либо отсутствует
          $ref: "#/components/responses/badRequest"
        "200":
          description: Успешно получен результат проверки
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    inside:
                      type: boolean
...